User Todos Laravel API

+++ app/Http/Controllers/TodosController.php
@@ -14,7 +14,7 @@ class TodosController extends Controller
      */
     public function index()
     {
-        return Todo::all();
+        return Todo::where('user_id',auth()->user()->id)->get();
     }
 
     /**
@@ -30,7 +30,11 @@ class TodosController extends Controller
             'completed'=>'required|boolean'
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed,
+        ]);
 
         return response($todo, 201);
     }
@@ -44,6 +48,10 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id){
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'completed'=>'required|boolean'
@@ -60,7 +68,7 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
         return response('Updated', 200);
     }
@@ -80,12 +88,29 @@ class TodosController extends Controller
 
     public function destroyCompleted(Request $request)
     {
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo){
+           return $todo->id;
+        });
+//collect 배열 데이터를 사용하기위한 유연하고 편리한 래퍼(wrapper)를 제공합니다.
+//every 메소드는 컬렉션의 모든 요소들이 전달된 조건을 충족하는지 확인하는데 사용할 수 있습니다:
+////contains 메소드는 컬렉션에 주어진 아이템을 포함하고 있는지 알려줍니다:
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds){
+           return $userTodoIds->contains($value);
+        });
+
+        if(!$valid){
+            return response()->json('Unauthorized', 401);
+        }
+
         $request->validate([
             'todos' => 'required|array',
         ]);
 
         Todo::destroy($request->todos);
 
-        return response('Deleted', 200);
+        return response()->json('Deleted', 200);
     }
+
 }
